+++ app/src/main/java/iss/nus/edu/medipalappln/application/App.java
@@ -1,20 +1,22 @@
-package iss.nus.edu.medipalappln.application;
-
-import android.app.Application;
-
-import iss.nus.edu.medipalappln.medipal.Club;
-
-
-/**
- * Created by swarna on 9/8/16.
- */
-public class App extends Application {
-  public static Club club;
-
-  @Override
-  public void onCreate() {
-    super.onCreate();
-    club = new Club();
-
-  }
-}
+//package iss.nus.edu.medipalappln.application;
+//
+//import android.app.Application;
+//import android.util.Log;
+//
+//import iss.nus.edu.medipalappln.medipal.Club;
+//
+//
+///**
+// * Created by swarna on 9/8/16.
+// */
+//public class App extends Application {
+//  public static Club club;
+//
+//  @Override
+//  public void onCreate() {
+//    super.onCreate();
+//    club = new Club();
+//    Log.i("APP", "onCreate: club created");
+//
+//  }
+//}
